feat: dataloader config - added for clarity and to separate custom configs from torch/transformers native configs
fix: train config
- fixed issue related to unsupported mixed precision type

style: removed unused lines from base dataset

fix: binary align models
- corrected issue with parsing configs to models

fix: binary align trainer
- added dataloader configs
- corrected issue with progress bar updating and steps calculation
- corrected parsing of collate function

+++ src/configs/train_config.py
@@ -9,7 +9,7 @@ from easydict import EasyDict
 class TrainConfig(ABC):
     experiment_name: str
     num_train_epochs: int = 5
-    mixed_precision: str = "fp32"
+    mixed_precision: str = "fp16"
     log_with: Optional[str] = "wandb"
     logging_steps: int = 200
     save_strategy: str = "epoch"
